workspacekit: Move rootfs-related code under pkg.
